+++ src/workerd/api/streams/common.h
@@ -25,7 +25,6 @@ class ReadableStreamController;
 class ReadableStreamSource;
 class ReadableStreamDefaultController;
 class ReadableByteStreamController;
-class ReadableStreamSource;
 
 class WritableStream;
 class WritableStreamController;
@@ -160,6 +159,85 @@ struct Transformer {
     transform?: (chunk: I, controller: TransformStreamDefaultController<O>) => void | Promise<void>;
     flush?: (controller: TransformStreamDefaultController<O>) => void | Promise<void>;
   });
+};
+
+// ReadableStreamSource and WritableStreamSink
+//
+// These are implementation interfaces for ReadableStream and WritableStream. If you just need to
+// use a ReadableStream or WritableStream, you can safely skip reading this. If you need to
+// implement a new kind of stream, read on.
+
+// In the original Workers streams implementation, a ReadableStream would have a
+// ReadableStreamSource backing it. Likewise, a WritableStream would have a WritableStreamSink.
+// The ReadableStreamSource and WritableStreamSink are kj heap objects that provide a thin
+// wrapper on internal native stream sources originating from within the Workers runtime.
+//
+// With implementation of full streams standard support, we introduce the new abstraction APIs
+// ReadableStreamController and WritableStreamController, which will provide the underlying
+// implementation for both ReadableStream and WritableStream, respectively.
+//
+// When creating a new kind of *internal* ReadableStream, where the data is originating internally
+// from a kj stream, you will still implement the ReadableStreamSource API, just as before.
+// Likewise, when creating a new kind of *internal* WritableStream, where the data destination is
+// a kj stream, you will implement the WritableStreamSink API.
+
+class WritableStreamSink {
+public:
+  virtual kj::Promise<void> write(const void* buffer, size_t size)
+      KJ_WARN_UNUSED_RESULT = 0;
+  virtual kj::Promise<void> write(kj::ArrayPtr<const kj::ArrayPtr<const byte>> pieces)
+      KJ_WARN_UNUSED_RESULT = 0;
+
+  virtual kj::Promise<void> end() KJ_WARN_UNUSED_RESULT = 0;
+  // Must call to flush and finish the stream.
+
+  virtual kj::Maybe<kj::Promise<DeferredProxy<void>>> tryPumpFrom(
+      ReadableStreamSource& input, bool end);
+
+  virtual void abort(kj::Exception reason) = 0;
+  // TODO(conform): abort() should return a promise after which closed fulfillers should be
+  //   rejected. This may necessitate an "erroring" state.
+};
+
+class ReadableStreamSource {
+public:
+  virtual kj::Promise<size_t> tryRead(void* buffer, size_t minBytes, size_t maxBytes) = 0;
+
+  virtual kj::Promise<DeferredProxy<void>> pumpTo(WritableStreamSink& output, bool end);
+  // The ReadableStreamSource version of pumpTo() has no `amount` parameter, since the Streams spec
+  // only defines pumping everything.
+  //
+  // If `end` is true, then `output.end()` will be called after pumping. Note that it's especially
+  // important to take advantage of this when using deferred proxying since calling `end()`
+  // directly might attempt to use the `IoContext` to call `registerPendingEvent()`.
+
+  virtual kj::Maybe<uint64_t> tryGetLength(StreamEncoding encoding);
+
+  kj::Promise<kj::Array<byte>> readAllBytes(uint64_t limit);
+  kj::Promise<kj::String> readAllText(uint64_t limit);
+
+  virtual void cancel(kj::Exception reason);
+  // Hook to inform this ReadableStreamSource that the ReadableStream has been canceled. This only
+  // really means anything to TransformStreams, which are supposed to propagate the error to the
+  // writable side, and custom ReadableStreams, which we don't implement yet.
+  //
+  // NOTE: By "propagate the error back to the writable stream", I mean: if the WritableStream is in
+  //   the Writable state, set it to the Errored state and reject its closed fulfiller with
+  //   `reason`. I'm not sure how I'm going to do this yet.
+  //
+  // TODO(conform): Should return promise.
+  //
+  // TODO(conform): `reason` should be allowed to be any JS value, and not just an exception.
+  //   That is, something silly like `stream.cancel(42)` should be allowed and trigger a
+  //   rejection with the integer `42`.
+
+  struct Tee {
+    kj::Own<ReadableStreamSource> branches[2];
+  };
+
+  virtual kj::Maybe<Tee> tryTee(uint64_t limit);
+  // Implement this if your ReadableStreamSource has a better way to tee a stream than the naive
+  // method, which relies upon `tryRead()`. The default implementation returns nullptr.
 };
 
 struct PipeToOptions {
@@ -591,6 +669,12 @@ public:
                      jsg::Optional<UnderlyingSink> underlyingSink,
                      jsg::Optional<StreamQueuingStrategy> queuingStrategy) {}
 };
+
+kj::Own<WritableStreamController> newWritableStreamJsController();
+kj::Own<WritableStreamController> newWritableStreamInternalController(
+    IoContext& ioContext,
+    kj::Own<WritableStreamSink> source,
+    kj::Maybe<uint64_t> maybeHighWaterMark = nullptr);
 
 struct Unlocked {};
 struct Locked {};
